### 2022-06-18 #4487 @arfon
@tonyewong – I believe the paper isn't compiling as Bakker's affiliations need quoting, i.e. like this:

```
  - name: Alexander Bakker
    orcid: 0000-0002-1017-7665
    affiliation: "5,7"
```
