### 2022-06-19 d5d43fb
Update paper.md
add quotations around multiple affiliations for Bakker
