### 2022-06-19 #4487 @tonyewong
https://github.com/openjournals/joss-reviews/issues/4487#issuecomment-1159554233

Sorry about that! Just updated paper.md to add the quotations.
