### 2022-06-29 #4487 @jbytecode
👋👋👋 Dear @lgloege 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**MimiBRICK.jl: A Julia package for the BRICK model for sea-level change in the Mimi integrated modeling framework**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4487).
