### 2022-07-04 #4487 @tpoisot
I took some time to go through the paper and repo - I have very minimal comments, and would gladly recommend @jbytecode to :tada: accept :tada: after the last little unchecked boxes below have been addressed (or discussed by the authors)

## Software license

- [x] the `LICENSE` file is there
- [ ] the package uses GNU, whereas most of the Julia ecosystem uses MIT - I was wondering if the authors could comment on this?

## Substantial scholarly effort

- [x] the (absurd) JOSS requirement on number of lines of code is met
- [x] there is a long commit history with several contributors
- [x] the framework has been cited, and the package is definitely setup in a way where it can have an impact (through citations, but also through adoption by stakeholders)

- [x] there is a documentation of the overall packages
- [x] there are vignettes as notebooks
- [x] there are docstrings for methods
- [ ] the `README` lacks a *statement of need* explaining what the package does to new users

## Installation instructions

- [x] installation instructions are given
- [ ] *however*, there has been no tagged release of the package, which I think should be corrected as it makes installation (and therefore use) easier

## Example usage

- [x] well covered -- far from my area of expertise but I thought it made sense

## API documentation

- [x] docstrings are attached to all the methods I checked
- [ ] *however*, I think the authors should give an overview of how pieces in the API fit together

- [x] adequate, but I encourage the authors to tick as many boxes on [this page](https://github.com/raddleverse/MimiBRICK.jl/community) as possible

- [x] I could install and run the examples

- [x] good coverage, the package uses CI
