### 2022-07-04 #4487 @jbytecode
@tpoisot - thank you for your quick review. It would be better to ask the steps rather than sending a reviewing report.

In the pre-review issues we always do initial stuff and find suitable reviewers. After starting the review, a new thread will be created. In this review thread, a checklist with 20 reviewing items for each reviewer. The process is interactive, that is, the author(s), the reviewers, and the editor always interact, reviewers are welcome to open new issues in the target repo, send pull requests, and any contributions are welcome.
