### 2022-07-06 #4487 @tonyewong
Sure, here are a few more. Again, I started from the bottom of the list and am working my way up. Generally looking for a combination of Julia and geo-/Earth sciences interests:
    * diegomcarvalho
    * zhenwu0728
    * haoxx081github.umn.edu
    * Zitzeronion
    * ruillercoaster
    * tbsexton
    * AnantHariharan1996
    * matt-graham
    * BrandonEdwards
    * crvernon
    * patrickcgray
