### 2022-07-07 #4487 @Zitzeronion
Hey @jbytecode & @tonyewong  👋

To be transparent, I haven't done a single review for **JOSS** yet, so it would be my first review.
That said, if you think that I am qualified, I wouldn't mind reviewing the package.

Let me know what you think 😃
