### 2022-07-08 #4556 @editorialbot
[REVIEW]: MimiBRICK.jl: A Julia package for the BRICK model for sea-level change in the Mimi integrated modeling framework
**Submitting author:** @tonyewong (<a href="http://orcid.org/0000-0002-7304-3883">Anthony Wong</a>)
**Repository:** https://github.com/raddleverse/MimiBRICK.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
