### 2022-07-17 #4556 @Zitzeronion
Hey @jbytecode,

I am sorry for the delay, started yesterday reviewing the project.
So far so good, one thing that is somewhat conflicting to me is the **Contribution and authorship**.
To my understanding, @lrennels (Lisa Rennels) is the creator and main contributor of the project.
@tonyewong contributed, but it is hard for me to assess his commits as major contributions.
Can @tonyewong please clarify this?
