### 2022-07-17 #4556 @jbytecode
Thank you @Zitzeronion for stressing this point.

- It seems the initial state of the model is published elsewhere: [Journal Article](https://gmd.copernicus.org/articles/10/2741/2017/)
- The first author is @tonyewong but they stated that the authors contributed equally to this work
- The abstract states that the initial attempts are written in R and FORTRAN.
- Does the role of @lrennels here consist of translating the original one into Julia? Why isn't @lrennels submitting autor?
- @tonyewong please clarify the issues that @Zitzeronion states.

Addition to this, the package is not installable using the standard way of Julia, e.g., `Pkg> add package`.  Although it is not a rigid requirement, as JOSS editors, we always encourage the authors to register their packages into the proper package server.
