### 2022-07-17 #4556 @tonyewong
@Zitzeronion @jbytecode 's assessment is accurate. I'm responsible for actively guiding/managing the project and doing some of the code development in tandem with @lrennels, who as you note is most active as a developer in terms of numbers of lines of code and commits.

Our team's understanding from the [JOSS submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) is that "active project direction" constitutes coauthor-worthy contributions to the project, and that the submitting author "must be _a_ major contributor to the software" (emphasis added). But we did not see a rule that the submitting/first author must be the one who has written the most lines of code for this repository, or specific guidelines about how active project management/direction should be evaluated vs. lines of code or numbers of commits. If there is a requirement, please let us know and we will happily adjust the authorship. If not, all coauthors have agreed to the coauthor ordering and roles.

We will add the MimiBRICK.jl package to the general package registry. @davidanthoff I think you had started to do this - can you check into adding the package?
