### 2022-07-17 #4556 @Zitzeronion
Thanks @jbytecode for putting such a nice bullet point list in such a short time.
Thanks @tonyewong for the answer, maybe one additional question is @Irennels a PhD student under your supervision?

> But we did not see a rule that the submitting/first author must be the one who has written the most lines of code for this repository, or specific guidelines about how active project management/direction should be evaluated vs. lines of code or numbers of commits.

I understand this reasoning, and I understand @tonyewong's role as active_guiding/managing_ of the project. My only concern is that this statement can't be easily seen from the repositories' metadata.
