### 2022-07-18 b408197
Update paper.md
Minor edits, doubling of "coupled BRICK"
