+++ joss_submission/paper.md
@@ -58,7 +58,7 @@ This implementation includes examples for using observational data to calibrate 
 
 BRICK is a semi-empirical model for global and local mean sea-level change [@Wong2017_brick]. The core model includes component sub-models for the major contributors to global mean sea-level change - glaciers and ice caps, thermal expansion, land water storage, and the Greenland and Antarctic ice sheets. The resulting global mean sea levels can be downscaled via a data set that represents the "fingerprint" of each sea-level component on local mean sea level [@slangen2014]. In this way, BRICK provides information about local sea-level changes, including characterizations of key uncertainties. BRICK is flexible and efficient enough to resolve high-risk upper tails of probability distributions. BRICK has been used in a number of recent assessments, including for examining the impacts of sea-level rise as a constraint on estimates of climate sensitivity [@Vega-Westhoff2018], estimates of deep uncertainty in coastal flood risk [@Ruckert2019], and most recently was included in comparisons of sea-level projections in the Sixth Assessment Report of the Intergovernmental Panel on Climate Change [@ipccar6ch9].
 
-By working with annual global mean temperatures and sea levels, BRICK is suitable for embedding within, and coupling to, other models for climate change and its impacts. `MimiBRICK.jl` is written in compliance with the Mimi integrated modeling framework to facilitate incorporating BRICK into larger-scale coupled modeling efforts. The `MimiBRICK.jl` repository includes three such examples: (i) standalone BRICK, which takes as input temperature and ocean heat uptake; (ii) BRICK coupled to a simple one-dimensional Diffusion-Ocean-Energy Climate model (DOECLIM) coupled to BRICK, which takes as input radiative forcing scenarios such as the standard Representative Concentration Pathway (RCP) scenarios; and (iii) BRICK coupled to a Simple Nonlinear Earth System model (SNEASY) coupled to BRICK, which takes as input radiative forcing and greenhouse gas emissions and concentration scenarios (such as the RCP scenarios).
+By working with annual global mean temperatures and sea levels, BRICK is suitable for embedding within, and coupling to, other models for climate change and its impacts. `MimiBRICK.jl` is written in compliance with the Mimi integrated modeling framework to facilitate incorporating BRICK into larger-scale coupled modeling efforts. The `MimiBRICK.jl` repository includes three such examples: (i) standalone BRICK, which takes as input temperature and ocean heat uptake; (ii) BRICK coupled to a simple one-dimensional Diffusion-Ocean-Energy Climate model (DOECLIM), which takes as input radiative forcing scenarios such as the standard Representative Concentration Pathway (RCP) scenarios; and (iii) BRICK coupled to a Simple Nonlinear Earth System model (SNEASY), which takes as input radiative forcing and greenhouse gas emissions and concentration scenarios (such as the RCP scenarios).
 
 The standalone BRICK model requires as input annual mean time series for global mean surface temperature and oceanic heat uptake. In the DOECLIM-BRICK and SNEASY-BRICK configurations, those temperatures and ocean heat uptake inputs are provided to BRICK through output from the DOECLIM and SNEASY models. In the examples provided in the `MimiBRICK.jl` repository, the temperature time series is from the maximum _a posteriori_ simulation from the ensemble run using the SNEASY-BRICK configuration. In all Mimi modeling experiments, model output and parameter values can be explored using the Mimi `explore()` function (\autoref{fig:explorer}). The `explore()` function allows users to easily view and zoom in on different features in the model simulation set up or results. Being coded in the Mimi style enables the user to more easily couple BRICK to the suite of other models already implemented in Mimi, and builds on the extensive documentation and community support online (https://www.mimiframework.org/Mimi.jl/stable/).
 
