### 2022-07-18 #4556 @tonyewong
Thanks for the clarification!

@Zitzeronion , would a suitable solution be to add a section to the repository README.md file that outlines Author Contributions? Or perhaps more broadly a "Contributions" section where we can give credit to our various users/contributors?
