### 2022-07-18 raddleverse/MimiBRICK.jl#52 @Zitzeronion
JOSS paper: State of the field 
@tonyewong  In the checklist for the paper appears the point:

- [x] State of the field

To check this off, I would like to ask you to add a section which compares MimiBRICK.jl to other software?
I assume that there are other probabilistic climate models even in Julia, but probably as well in Python.
Is there a quasi standard package that more than 10 different academic groups use, or does every group create their own project?
