### 2022-07-18 #4556 @jbytecode
The review threads are always attached to the published papers, so what we discussed here is available for those who are interested in.

We can leave this issue as is, or, of course it can be added as a small note in the paper as well as in the README.md.
