+++ Project.toml
@@ -1,50 +1,50 @@
 name = "MimiBRICK"
 uuid = "91f4afc7-28cd-4627-b11a-a83dc9ff1a58"
-version = "1.0.0-DEV"
+version = "1.0.0"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-CSVFiles = "5d742f6a-9f54-50ce-8119-2520741973ca"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
-KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 MCMCDiagnostics = "6e857e4b-079a-58c4-aeab-bc2670384359"
-Mimi = "e4e893b0-ee5e-52ea-8111-44b3bdec128c"
-MimiSNEASY = "3f1801d0-c5f8-11e9-1568-bf13fd643260"
-Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
-NetCDF = "30363a11-5582-574a-97bb-aa9a979735b9"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Query = "1a8c2f83-1ff3-5112-b086-8aa67b057ba1"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+CSVFiles = "5d742f6a-9f54-50ce-8119-2520741973ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 RobustAdaptiveMetropolisSampler = "2f96e190-b8a6-11e9-0b3d-5fbd22c21613"
-SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
+Query = "1a8c2f83-1ff3-5112-b086-8aa67b057ba1"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
+MimiSNEASY = "3f1801d0-c5f8-11e9-1568-bf13fd643260"
+KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
-
-[compat]
-CSV = "0.9, 0.10"
-CSVFiles = "1"
-DataFrames = "1"
-Distributions = "0.23,0.24,0.25"
-Interpolations = "0.13"
-KernelDensity = "0.6"
-MCMCDiagnostics = "0.3"
-Mimi = "1.2"
-MimiSNEASY = "1"
-Missings = "1"
-NetCDF = "0.11"
-Plots = "1"
-Query = "1"
-RobustAdaptiveMetropolisSampler = "1"
-SpecialFunctions = "1, 2"
-StatsBase = "0.33"
-julia = "1.6"
+Mimi = "e4e893b0-ee5e-52ea-8111-44b3bdec128c"
[…]
