### 2022-07-19 #4556 @davidanthoff
I just submitted the package to the General Julia registry, see https://github.com/JuliaRegistries/General/pull/64547. It should get merged in about three days, which is currently the wait time for new package registrations.
