+++ joss_submission/paper.bib
@@ -1,3 +1,39 @@
+@software{garner2022,
+  author       = {Garner, Gregory G. and
+                  Kopp, Robert E.},
+  title        = {{Framework for Assessing Changes To Sea-level
+                   (FACTS) modules, scripts, and data for the IPCC
+                   AR6 sea level projections}},
+  month        = apr,
+  year         = 2022,
+  note         = {{The development of the sea-level rise projections
+                   was supported by multiple funders, including the
+                   U.S. National Aeronautics and Space Administration
+                   (grants 80NSSC17K0698, 80NSSC20K1724 and
+                   80NSSC21K0322 and JPL task
+                   105393.509496.02.08.13.31), the U.S. National
+                   Science Foundation (grant ICER-1663807), the U.K.
+                   Natural Environment Research Council (grant
+                   NE/T009381/1), NIOZ Royal Netherlands Institute
+                   for Sea Research, PROTECT (which has received
+                   funding from the European Union's Horizon 2020
+                   research and innovation programme under grant
+                   agreement no. 869304), and UK Natural Environment
+                   Research Council grant NE/T007443/1. We
+                   acknowledge the World Climate Research Programme,
+                   which, through its Working Group on Coupled
+                   Modelling, coordinated and promoted CMIP6. We
+                   thank the climate modeling groups for producing
+                   and making available their model output, the Earth
+                   System Grid Federation (ESGF) for archiving the
+                   data and providing access, and the multiple
+                   funding agencies who support CMIP6 and ESGF.}},
+  publisher    = {Zenodo},
+  version      = 20220406,
+  doi          = {10.5281/zenodo.6419954},
+  url          = {https://doi.org/10.5281/zenodo.6419954}
+}
+
 @article{slangen2014,
 author = {Slangen, A B A and Carson, M and Katsman, C A and van de Wal, R S W and K{\"{o}}hl, A and Vermeersen, L L A and Stammer, D},
 doi = {10.1007/s10584-014-1080-9},
