+++ joss_submission/paper.md
@@ -50,13 +50,13 @@ bibliography: paper.bib
 
 # Statement of need
 
-Assessing strategies to manage climate risks requires sound projections of future climate. For coastal risks, this includes projections of future sea levels. Major contributors to sea-level change include glaciers and ice caps, thermal expansion, land water storage, and the Greenland and Antarctic ice sheets. Characterizing coastal hazards and managing the associated risks requires resolving the tails of distributions. Semi-empirical models for sea-level rise offer a computationally efficient method for characterizing uncertainties in future coastal hazards and fusing observational data with models [@Kopp2017; @Mengel2016; @Nauels2017; @Wong2017_brick]. These models can also be flexible and modular, which enables their use in integrated frameworks for assessing climate damages and examining the efficacy of climate risk management policies. The Mimi integrated modeling framework (https://www.mimiframework.org/) is a coding platform that facilitates coupling models and running coupled modeling experiments.`MimiBRICK.jl` is an implementation of the Building Blocks for Relevant Ice and Climate Knowledge (BRICK) semi-empirical model for sea-level change [@Wong2017_brick] in the Mimi framework. `MimiBRICK.jl` is flexible, efficient, and modular, to facilitate incorporating BRICK into coupled models and integrated assessments of climate impacts in a modular fashion to provide global average as well as local sea-level projections [@national_academies_of_sciences_valuing_2017].
+Assessing strategies to manage climate risks requires sound projections of future climate. For coastal risks, this includes projections of future sea levels. Major contributors to sea-level change include glaciers and ice caps, thermal expansion, land water storage, and the Greenland and Antarctic ice sheets. Characterizing coastal hazards and managing the associated risks requires resolving the tails of distributions. While many practitioners can obtain sea-level scenarios from syntheses reports, including the Intergovernmental Panel on Climate Change (IPCC, [@ipccar6ch9]), improving our understanding of the importance and impacts of the many inherent uncertainties often requires fast and efficient models to explore these uncertainties. Semi-empirical models for sea-level rise offer a computationally efficient method for studying uncertainties in future coastal hazards and fusing observational data with models [@Kopp2017; @Mengel2016; @Nauels2017; @Wong2017_brick]. These models can also be flexible and modular, which enables their use in integrated frameworks for assessing climate damages and examining the efficacy of climate risk management policies. Interfaces such as Framework for Assessing Changes To Sea-level (FACTS, [@garner2022]) enable users to access these models and facilitate experimentation and intercomparisons. Along this line, the Mimi integrated modeling framework (https://www.mimiframework.org/) is a coding platform that facilitates coupling models and running coupled modeling experiments. Mimi is explicitly employed in more than 10 peer-reviewed studies, as of this writing, with more presently in review. `MimiBRICK.jl` is an implementation of the Building Blocks for Relevant Ice and Climate Knowledge (BRICK) semi-empirical model for sea-level change [@Wong2017_brick] in the Mimi framework. `MimiBRICK.jl` is flexible, efficient, and modular, to facilitate incorporating BRICK into coupled models and integrated assessments of climate impacts in a modular fashion to provide global average as well as local sea-level projections [@national_academies_of_sciences_valuing_2017].
 
 This implementation includes examples for using observational data to calibrate the model, as well as various configurations in which `MimiBRICK` is coupled to other climate model components. For users who do not wish to re-run computationally intensive model calibration algorithms, this implementation also includes scripts for using existing calibration output for standard future climate change scenarios, and examples downscaling these global projections for assessments of local impacts.
 
 # Summary
 
[…]
