### 2022-07-19 raddleverse/MimiBRICK.jl#52 @tonyewong
Check out the PR here: https://github.com/raddleverse/MimiBRICK.jl/pull/54#issue-1309801567
and let me know if there are further edits you'd suggest. We're being a bit cautious not to add too much to the text to stay within the 250-1000 word JOSS guidelines. Thanks!
