### 2022-07-19 #4556 @tonyewong
@Zitzeronion @jbytecode I went ahead and added a brief Author Contributions section to the top-level README.md.

https://github.com/raddleverse/MimiBRICK.jl/pull/54#issue-1309801567
