### 2022-07-20 #4556 @Zitzeronion
I've been active in the repo.
There is just one thing left from the checkbox for the paper, which is **State of the field**.

The paper is well written, but I've hoped to see at least one comparison with another OSS package.
