### 2022-07-21 85d37f6
adding more details on state of field
