+++ joss_submission/paper.bib
@@ -1,3 +1,38 @@
+@misc{srikrishnan_2022,
+	type = {preprint},
+	title = {Uncertainty analysis in multi-sector systems: {Considerations} for risk analysis, projection, and planning for complex systems},
+	copyright = {Creative Commons Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)},
+	shorttitle = {Uncertainty analysis in multi-sector systems},
+	url = {http://www.essoar.org/doi/10.1002/essoar.10510113.2},
+	abstract = {Simulation models of multi-sector systems are increasingly used to understand societal resilience to climate and economic shocks and change. However, multi-sector systems are also subject to numerous},
+	language = {EN},
+	urldate = {2022-06-03},
+	journal = {Earth and Space Science Open Archive},
+	author = {Srikrishnan, Vivek and Lafferty, David C. and Wong, Tony E. and Lamontagne, Jonathan R. and Quinn, Julianne D. and Sharma, Sanjib and Molla, Nusrat J. and Herman, Jonathan D. and Sriver, Ryan L. and Morris, Jennifer and Lee, Ben Seiyon},
+	month = may,
+	year = {2022},
+	doi = {10.1002/essoar.10510113.2},
+	note = {Archive Location: world
+Publisher: Earth and Space Science Open Archive
+Section: Climatology (Global Change)},
+	file = {Snapshot:/Users/aewsma/Zotero/storage/QDNL89YJ/essoar.10510113.html:text/html},
+}
+
+@article{kopp_2014,
+	title = {Probabilistic 21st and 22nd century sea-level projections at a global network of tide-gauge sites},
+	volume = {2},
+	issn = {2328-4277},
+	url = {http://dx.doi.org/10.1002/2014EF000239},
+	doi = {10.1002/2014EF000239},
+	abstract = {Sea-level rise due to both climate change and non-climatic factors threatens coastal settlements, infrastructure, and ecosystems. Projections of mean global sea-level (GSL) rise provide insufficient information to plan adaptive responses; local decisions require local projections that accommodate different risk tolerances and time frames and that can be linked to storm surge projections. Here we present a global set of local sea-level (LSL) projections to inform decisions on timescales ranging from the coming decades through the 22nd century. We provide complete probability distributions, informed by a combination of expert community assessment, expert elicitation, and process modeling. Between the years 2000 and 2100, we project a very likely (90\% probability) GSL rise of 0.5–1.2 m under representative concentration pathway (RCP) 8.5, 0.4–0.9 m under RCP 4.5, and 0.3–0.8 m under RCP 2.6. Site-to-site differences in LSL projections are due to varying non-climatic background uplift or subsidence, oceanographic effects, and spatially variable responses of the geoid and the lithosphere to shrinking land ice. The Antarctic ice sheet (AIS) constitutes a growing share of variance in GSL and LSL projections. In the global average and at many locations, it is the dominant source of variance in late 21st century projections, though at some sites oceanographic processes contribute the largest share throughout the century. LSL rise dramatically reshapes flood risk, greatly increasing the expected number of “1-in-10” and “1-in-100” year events.},
+	number = {8},
+	journal = {Earth's Future},
+	author = {Kopp, Robert E and Horton, Radley M and Little, Christopher M and Mitrovica, Jerry X and Oppenheimer, Michael and Rasmussen, D J and Strauss, Benjamin H and Tebaldi, Claudia},
+	year = {2014},
+	note = {ISBN: 2328-4277},
+	keywords = {climate change, sea level, coastal flooding, 1641 Sea level change, uncertainty quantification, 1821 Floods, risk analysis},
+}
+
 @software{garner2022,
   author       = {Garner, Gregory G. and
                   Kopp, Robert E.},
