### 2022-07-21 #4556 @tonyewong
To keep @jbytecode updated - The comparison was added to the paper, and @Zitzeronion has closed the issue (https://github.com/raddleverse/MimiBRICK.jl/issues/52). Thanks for the useful feedback @Zitzeronion !
