### 2022-07-31 raddleverse/MimiBRICK.jl#57 @tonyewong
package-ify 
Swap the .git references in installation/loading to proper package references
https://github.com/JuliaRegistries/General/tree/master/M/MimiBRICK
