### 2022-08-08 raddleverse/MimiBRICK.jl#57 @lrennels
#58 should handle this @tonyewong
