### 2022-08-08 #4556 @tonyewong
@jbytecode I'll also mention that we have just merged a PR that uses MimiBRICK.jl now as a proper Julia package, instead of referring to the .git repo.
