### 2022-08-08 #4556 @jbytecode
@tonyewong - thank you for informing us. I have just installed the package in the standard way and run a full test. It seems it is okay as it passes all of the test cases.

one more thing about tests is you can use testset macros like (even though it is not mandatory)

```julia
@testset "Section" begin
    @testset "Subsection" begin
         @test expected == actual
    end
end
```

here is an example: [use of testset macro](https://github.com/jbytecode/JMcDM/blob/main/test/testmcdm.jl)

Another issue is the structure of the text. The summary part comes last. Maybe we can refactor the text after our reviewers make a decision based on their reviews.
