### 2022-08-08 #4556 @lrennels
@jbytecode sounds good, I think we've done so to a certain extent in the master `runtests.jl` file [here](https://github.com/raddleverse/MimiBRICK.jl/blob/master/test/runtests.j) but could certainly add more hierarchy to it.
