### 2022-08-09 #4556 @tonyewong
We structured the paper with the Statement of Need first and the Summary second, to first present a need, and then describe how the package addresses this need. We of course noted that most JOSS papers have the Summary first, but we saw that some JOSS papers follow the structure with the Summary following the Statement of Need (e.g., [this one](https://joss.theoj.org/papers/10.21105/joss.04298) or [this one](https://joss.theoj.org/papers/10.21105/joss.04266)) and we liked that logical flow. While it isn't our preference, we are happy to refactor the paper, if needed, based on reviewer and editorial preferences.
