### 2022-08-12 #4556 @Zitzeronion
Yesterday I worked on the remaining points @jbytecode. For the release of the JOSS paper I thought it would be nice to get rid of the warnings during testing, so I created an issue and a PR with a possible solution.
Overall I think the package is good and has a clear scientific use case.
Minor criticism:
1. Function style could be improved
For example in *BRICK_DOECLIM.jl* from line 99 to line 122 there are 23 consecutive calls of *update_param!()*.
Nothing wrong with that, I just think that a loop over the last two arguments would be clearer.
2. *test_full_run.jl* should be dropped
Correct me if I'm wrong here, but all functionality is already tested in *test_downscaling.jl*.
I do understand the reasoning, but a test file that hasn't a single test statement is somewhat strange.
3. Community guidelines
Do the authors (@tonyewong) want to add Community guidelines in their *Readme.md*? I just want to ask, a no is fine with me and I will check the box.
4. Documentation/Pages
If it isn't too much work, I would really appreciate it if the authors could make use of github-pages. A docs folder could help to show use cases, with code and plots, similar to [Mimi](https://www.mimiframework.org/). It would help to reduce the amount of the text in the *Readme.md* because information could be offloaded to the first documentation page and it would create a place to collect all the docstrings.

I recommend the article for publication.
