### 2022-08-13 #4556 @tonyewong
I suspect @tpoisot is simply [still offline](https://github.com/openjournals/joss-reviews/issues/4487#issuecomment-1174068872).

I'm happy to either use their [previous post](https://github.com/openjournals/joss-reviews/issues/4487#issuecomment-1174056698) in the pre-review issue and address the points they raised there, or to wait to hear from them when they return, or to hunt for an alternative reviewer. Of course, as they are already familiar with the submission there are benefits to keeping them on.
