### 2022-08-14 #4556 @tonyewong
Addressing some of the open issues in @tpoisot 's [original review post](https://github.com/openjournals/joss-reviews/issues/4487#issuecomment-1174056698)

>  the package uses GNU, whereas most of the Julia ecosystem uses MIT - I was wondering if the authors could comment on this?

The previous iterations of the BRICK code, on which this is built, use a GNU license. Others among the Mimi models with which we anticipate MimiBRICK will be used also use a GNU license (e.g., [MimiSNEASY](https://github.com/anthofflab/MimiSNEASY.jl/blob/master/LICENSE)). So we figured staying consistent with that license is advantageous.

>  however, there has been no tagged release of the package, which I think should be corrected as it makes installation (and therefore use) easier

I believe the now-closed [issue 57](https://github.com/raddleverse/MimiBRICK.jl/issues/57) in the repository addressed this.

Open issues that I will create an issue in the repo to address, alongside the issue opened by @Zitzeronion , so they're all documented in one place:

> the README lacks a statement of need explaining what the package does to new users
>
>  however, I think the authors should give an overview of how pieces in the API fit together
