### 2022-08-14 #4556 @jbytecode
@tonyewong - did you literally copy-pasted some parts of a legacy GNU-licensed code? If it is not the case, this is a Julia implementation and it can be licensed under MIT, imo. I am agree with the reviewer's suggestions here since most of the Julia packages are licensed under MIT.
