### 2022-08-15 #4556 @tonyewong
Yes, there are sections where some legacy code, which was under a GNU license, was copy-pasted and translated to Julia. We very much appreciate the simplicity and popularity of the MIT license. However, given that the [GNU GPLv3](https://www.gnu.org/licenses/gpl-3.0.en.html) says the following about modified versions of the code, we are most comfortable releasing software containing modifications of GPLv3-licensed code:

> You may convey a work based on the Program, or the modifications to produce it from the Program, in the form of source code under the terms of section 4, provided that you also meet all of these conditions:
> ...
> c) You must license the entire work, as a whole, under this License to anyone who comes into possession of a copy. This License will therefore apply, along with any applicable section 7 additional terms, to the whole of the work, and all its parts, regardless of how they are packaged. This License gives no permission to license the work in any other way, but it does not invalidate such permission if you have separately received it.
