### 2022-08-15 84b6ffb
Update README.md
moving examples to github pages, out of readme.md
