+++ README.md
@@ -25,102 +25,9 @@ This implementation includes examples for using observational data to calibrate 
   * Please let us know of any new results that should be incorporated into the [Introduction-and-Library repository](https://github.com/raddleverse/Introduction-and-Library)! See the README.md file in that repository for more information.
   * Also please let us know of any model modifications and/or bug fixes that might usefully be incorporated into the main BRICK codes. Creating an "Issue" here is a great way to do that (top horizontal menu bar in the GitHub browser).
 
-## How To Install MimiBRICK
+## Installation and Examples
 
-This code was created using [Julia v1.6](https://julialang.org/downloads/) and requires several Julia packages. It is recommended that you use Julia v1.6 (or later). Julia may be downloaded from http://julialang.org/downloads/.
-
-(1) Run the following line to install the Mimi implementation of BRICK:
-
-```julia
-]
-add MimiBRICK
-```
-
-_Note: Once MimiBRICK is officially published and packaged, this will be replaced by a simpler `add MimiBRICK`. But until that time, just point at the Github URL there._
-
-(2) To exit back to Julia, hit the `backspace` key.
-
-<br>
-
-## Running baseline cases with default parameters and unit tests
-
-If you would like to take a look at the unit tests run with Continuous Integration for this package, feel free to take a look at the `runtests.jl` file, and the separate testing scripts it calls [here](https://github.com/raddleverse/MimiBRICK.jl/tree/master/test).
-
-### More detailed examples
-
-Below, we review the three main configurations of the model that we anticipate being used. However, in the `examples` directory, you can find further examples conducting the statistical calibration, creating posterior model hindcasts and projections, and downscaling these hindcasts and projections to local sea-level changes. Those examples are in Jupyter notebooks. To open and run one, you will need to clone or fork the MimiBRICK.jl repository. Then, navigate to it in the Julia terminal and run:
-```julia
-using Pkg; Pkg.add(IJulia)
-using IJulia; notebook(dir="examples")
-```
-You can then select and run the Jupyter notebook of your chosing in the Jupyter browser window (which will open in a web browser).
-
-### BRICK standalone (with temperature and ocean heat uptake exogenous forcing)
-
-This is the first test that is done in `test/runtests/test_default.jl`. Since it does not require DOECLIM or SNEASY, you can run BRICK using temperature and ocean heat uptake forcing data by running in the Julia console:
-```julia
-using MimiBRICK
-m = MimiBRICK.get_model()
-run(m)
-```
-
-You can plot the output fields in the model object `m` using (for example) the `Plots` Julia plotting package. First, let's grab the years over which the model was run. This is a dimension in the model. We can retrieve it by using the `dim_keys()` function, from the `Mimi` package.
-```julia
-using Mimi
-years = dim_keys(m, :time)
-```
-
-Then we can load the `Plots` package and make a figure of the global mean sea-level change. Note that the first argument into the `m` object specifies the component of our model, and the second argument specifies the field. Here, we are grabbing the `sea_level_rise` field from the `global_sea_level` component.
-```julia
-using Plots
-plot(years, m[:global_sea_level, :sea_level_rise])
-```
-
-Mimi also offers an explorer window to check these model output fields out. To use this, we need to load the `Mimi` package (if you haven't already).
-```julia
-using Mimi
-```
-
-Then, we can open the explorer.
-```julia
-explore(m)
-```
-
[…]
