### 2022-08-15 3cd442b
Delete test_full_run.jl
per @Zitzeronion 's comment; this is fully included in test_downscale
