### 2022-08-15 8e1b523
Update README.md
updating Contributions section to be more general
