### 2022-08-15 #4556 @tonyewong
Re: @Zitzeronion 's comments [here](https://github.com/openjournals/joss-reviews/issues/4556#issuecomment-1212878859)

> 1. Function style could be improved
>    For example in BRICK_DOECLIM.jl from line 99 to line 122 there are 23 consecutive calls of update_param!().
>    Nothing wrong with that, I just think that a loop over the last two arguments would be clearer.
>    test_full_run.jl should be dropped

Thanks for the suggestion! Yes, this is definitely able to be written much more compactly. We will work this in where appropriate, but for now we prefer to leave all of the parameters explicitly set since they're model parameters that someone might be interested in manipulated "by hand".

>2.    Correct me if I'm wrong here, but all functionality is already tested in test_downscaling.jl.
>    I do understand the reasoning, but a test file that hasn't a single test statement is somewhat strange.

Great point - this has been [removed from the repo](https://github.com/raddleverse/MimiBRICK.jl/commit/3cd442b4deadac1c2b4150204cf3ed1d4cd8eae5).

>3.    Community guidelines
>    Do the authors (@tonyewong) want to add Community guidelines in their Readme.md? I just want to ask, a no is fine with me and I will check the box.

For now, we will leave out an explicit MimiBRICK statement, but have [modified](https://github.com/raddleverse/MimiBRICK.jl/commit/8e1b523a22f7d9d2a346db49cb0e370837d50e87) the Author Contributions section of the README to be more general to include general users. This points users to how the preferred methods for engaging with us and contributing are via GitHub (eg, issues, PRs) or the Mimi Framework Forum.

> 4.   Documentation/Pages
>    If it isn't too much work, I would really appreciate it if the authors could make use of github-pages. A docs folder could help to show use cases, with code and plots, similar to [Mimi](https://www.mimiframework.org/). It would help to reduce the amount of the text in the Readme.md because information could be offloaded to the first documentation page and it would create a place to collect all the docstrings.

Another great suggestion. I've enabled Github Pages for the project and moved the Installation and Examples from the top-level README.md into the Pages documentation. I'm sure I'll keep refactoring and reorganizing it, but for now that does tidy things up well.
