### 2022-08-16 #4556 @jbytecode
@svchb - here is the latest version of the paper.
