### 2022-08-16 dfd27bd
Update README.md
removing references to aspirational library of previously published results
