### 2022-08-16 #4556 @tonyewong
@jbytecode Am I correct that you are referring to the Community Guidelines point from @Zitzeronion 's checklist above? I think that it is now somewhat buried after the activity the past few days, but in a [later post](https://github.com/openjournals/joss-reviews/issues/4556#issuecomment-1214904486) I responded as:

> > Community guidelines
> >        Do the authors (@tonyewong) want to add Community guidelines in their Readme.md? I just want to ask, a no is fine with me and I will check the box.
>
> For now, we will leave out an explicit MimiBRICK statement, but have [modified](https://github.com/raddleverse/MimiBRICK.jl/commit/8e1b523a22f7d9d2a346db49cb0e370837d50e87) the Author Contributions section of the README to be more general to include general users. This points users to how the preferred methods for engaging with us and contributing are via GitHub (eg, issues, PRs) or the Mimi Framework Forum.

I'll let @Zitzeronion decide whether this satisfactorily addresses this last point. Happy to make further adjustments if needed!
