### 2022-08-16 #4556 @jbytecode
@tonyewong - it is usual to introduce the installation instructions in a top level of the README.md file of Julia packages like

```julia
julia> using Pkg
julia> Pkg.add(...)
```

I know, you have already provided this information in page https://raddleverse.github.io/MimiBRICK.jl/ but It would be nice to introduce the code snippet that I have given above in README.md.

Thank you in advance.
