+++ README.md
@@ -28,7 +28,16 @@ This implementation includes examples for using observational data to calibrate 
 
 ## Installation and Examples
 
-Installation instructions and example use cases can be found in the GitHub Pages site: https://raddleverse.github.io/MimiBRICK.jl/
+This code was created using Julia v1.6 and requires several Julia packages. It is recommended that you use Julia v1.6 (or later). Julia may be downloaded from http://julialang.org/downloads/.
+
+(1) Run the following line to install the Mimi implementation of BRICK:
+
+]
+add MimiBRICK
+
+(2) To exit back to Julia, hit the backspace key.
+
+Further instructions and example use cases can be found in the GitHub Pages site: https://raddleverse.github.io/MimiBRICK.jl/
 
 <br>
 
