### 2022-08-16 #4556 @tonyewong
Sure thing! I just added our brief section about Installation from the Github pages site. Thanks for the good suggestion!
