+++ docs/README.md
@@ -36,8 +36,6 @@ This code was created using [Julia v1.6](https://julialang.org/downloads/) and r
 add MimiBRICK
 ```
 
-_Note: Once MimiBRICK is officially published and packaged, this will be replaced by a simpler `add MimiBRICK`. But until that time, just point at the Github URL there._
-
 (2) To exit back to Julia, hit the `backspace` key.
 
 <br>
