### 2022-08-17 e33d1ba
Update README.md
adding description of the code structure
