### 2022-08-17 022163b
Update README.md
removing reference to aspirational/under construction library repository
