### 2022-08-17 #4556 @tonyewong
@svchb Thanks!  I just removed the reference to the aspirational library from the github pages page (good catch!), and I added a high-level overview of the structure of the repository to the top-level README.md (https://github.com/raddleverse/MimiBRICK.jl/commit/e33d1ba12e6f31e4623fddc050c167b0f9ca9d73)
