### 2022-08-17 #4556 @jbytecode
I think our reviewer ask you to give a small brief of explanation about the design pattern you used to develop your models.
