### 2022-08-17 #4556 @jbytecode
@tonyewong - please take a look at the section of `design pattern` in README.md [here](https://github.com/jbytecode/jmcdm). There is a clean description of development of a new method (how to create or extend data types, how to design functions, how to define return types, etc.)
