### 2022-08-17 #4556 @tonyewong
I was addressing this one:

> While your code is not that large (atm) it might still be nice to give some details on the structure of the code on your documentation page.

But I can certainly also add some text about the design pattern and how to add additional or substitute submodels.
