### 2022-08-17 3a2ef0d
Update README.md
adding more information about modifying component submodels and design pattern
