### 2022-08-17 #4556 @tonyewong
@jbytecode @svchb Okay, just added (https://github.com/raddleverse/MimiBRICK.jl/commit/3a2ef0df4e1c158a290ba76897277c49b0b6449a) a bit about the design pattern and what is involved in modifying/adding new component submodels. I tried not to be too verbose as the [Mimi Documentation](https://www.mimiframework.org/Mimi.jl/stable/) gives much more detailed instructions on adding/modifying/using models in the Mimi Framework. But, let me know if there are things that seem to be left out that would be useful to add.
