### 2022-08-17 #4556 @jbytecode
@tonyewong - It seems our reviewers are okay with the reviewing item of "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?"

It is very common in Julia packages to have an API documentation, especially generated by the Documenter.jl. Would you consider adding this kind of stuff for package? I think having that kind of documentation would increase the quality of your package.

Here is an example: https://gsa.sciml.ai/dev/
