### 2022-08-17 #4556 @tonyewong
@jbytecode Sure, I'm happy to work on adding further documentation via Documenter.jl. That's consistent with the parent Mimi.jl repository anyway (https://github.com/mimiframework/Mimi.jl/blob/master/docs). I can do this in parallel with your editorial work if that's okay with you.
