### 2022-08-17 #4556 @tonyewong
Looks great - thanks! Just merged it.

I'll create an issue in the repo for adding the Documenter.jl documentation. I completely agree that it will be a good enhancement for the package's documentation!

Many thanks @Zitzeronion and @svchb for the helpful comments and quick review! Many thanks as well to @jbytecode for the helpful feedback and managing the process!
