### 2022-08-17 #4556 @tonyewong
Sorry I may have misunderstood - Should I work on the Documenter.jl stuff as a condition for acceptance? Or as an enhancement to the documentation later, since the reviewers are happy with it as-is?
