+++ docs/index.md
@@ -1,0 +1,3 @@
+# MimiBRICK.jl
+
+placeholder
