+++ docs/index.md
@@ -1,3 +1,0 @@
-# MimiBRICK.jl
-
-placeholder
