+++ docs/src/installation.md
@@ -1,12 +1,0 @@
-# Installation
-
-This code was created using Julia v1.6 and requires several Julia packages. It is recommended that you use Julia v1.6 (or later). Julia may be downloaded from http://julialang.org/downloads/.
-
-(1) Run the following line to install the Mimi implementation of BRICK:
-
-```
-]
-add MimiBRICK
-```
-
-(2) To exit back to Julia, hit the backspace key.
