+++ docs/src/index.md
@@ -1,75 +1,33 @@
 # MimiBRICK.jl
 
-```@contents
+MimiBRICK.jl is an implementation of the [Building Blocks for Relevant Ice and Climate Knowledge (BRICK) semi-empirical model for sea-level change](https://doi.org/10.5194/gmd-10-2741-2017) in the Mimi integrated modeling framework (https://www.mimiframework.org/). The Mimi modeling framework is a coding platform that facilitates coupling models and running coupled modeling experiments. MimiBRICK.jl is flexible, efficient, and modular, to facilitate incorporating BRICK into coupled models and integrated assessments of climate impacts in a modular fashion to provide global average as well as local sea-level projections. This focus on tight model coupling and integrated modeling is a key feature of MimiBRICK.jl and broader Mimi modeling framework.
+
+This implementation includes examples for using observational data to calibrate the model, as well as various configurations in which MimiBRICK.jl is coupled to other climate model components. For users who do not wish to re-run computationally intensive model calibration algorithms, this implementation also includes scripts for using existing calibration output for standard future climate change scenarios, and examples downscaling these global projections for assessments of local impacts. Pre-run model calibration and simulation output can be found in the [accompanying Zenodo repository](https://zenodo.org/record/6626335).
+
+```@autodocs
+Modules = [MimiBRICK]
 ```
 
-### Functions for creating typical coupled model instances beyond stand-alone BRICK.
+<br>
 
-```@docs
+## Author Contributions
 
-MimiBRICK.create_brick_doeclim(;rcp_scenario::String = "RCP85", start_year::Int=1850, end_year::Int=2020)
+* TW: initial model development, software development, model calibration and validation, conceptualization, projection direction and overall management
+* LR: software development, package maintenance, conceptualization
+* FE: software development, model calibration and validation, conceptualization
+* VS: software testing, model calibration and validation, conceptualization
+* AB: initial model development, software testing, conceptualization
+* KK: software testing, conceptualization
+* DA: software development, package maintenance, conceptualization, project direction
 
-MimiBRICK.create_sneasy_brick(;rcp_scenario::String = "RCP85", start_year::Int=1850, end_year::Int=2020)
-```
+<br>
 
-### Functions for creating prior and posterior distributions for BRICK and typical coupled model configuration (including likelihoods)
+## License
 
-```@docs
-MimiBRICK.construct_brick_log_prior(joint_antarctic_prior::Bool; calibration_data_dir::Union{String, Nothing} = nothing)
+Copyright 2022 Tony Wong, Lisa Rennels, Frank Errickson, Vivek Srikrishnan, Alexander Bakker, Klaus Keller, and David Anthoff
 
-MimiBRICK.construct_brick_log_posterior(f_run_model!; model_start_year::Int=1852, calibration_end_year::Int=2017, joint_antarctic_prior::Bool=false)
+This file and codes in this repository are part of MimiBRICK.jl (Building blocks for Relevant Ice and Climate Knowledge). MimiBRICK.jl is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
 
-MimiBRICK.construct_doeclimbrick_log_prior(joint_antarctic_prior::Bool, uniform_ECS::Bool; calibration_data_dir::Union{String, Nothing} = nothing)
[…]
