### 2022-08-19 71e4939
Remove v1.8 testing for now.
