+++ .jlpkgbutler.toml
@@ -1,1 +1,2 @@
 strategy-matrix-exclude = "os: windows-latest, julia-arch: x86; os: ubuntu-latest, julia-arch: x86"
+strategy-matrix-julia-version = "['1.4', '1.5', '1.6', '1.7']"
