+++ .jlpkgbutler.toml
@@ -1,2 +1,1 @@
 strategy-matrix-exclude = "os: windows-latest, julia-arch: x86; os: ubuntu-latest, julia-arch: x86"
-strategy-matrix-julia-version = "['1.6', '1.7']"
