+++ README.md
@@ -1,4 +1,4 @@
-![](https://github.com/raddleverse/MimiBRICK.jl/workflows/jlpkgbutler-ci-master-workflow.yml/badge.svg)
+![](https://github.com/raddleverse/MimiBRICK.jl/actions/workflows/jlpkgbutler-ci-master-workflow.yml/badge.svg)
 [![codecov](https://codecov.io/gh/raddleverse/MimiBRICK.jl/branch/master/graph/badge.svg?token=H7SJB47W5V)](https://codecov.io/gh/raddleverse/MimiBRICK.jl)
 
 # MimiBRICK.jl
