+++ joss_submission/paper.bib
@@ -18,24 +18,21 @@ eprint = {
     abstract = { Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be “laws of nature" by practitioners of numerical computing: \beginlist \item High-level dynamic programs have to be slow. \item One must prototype in one language and then rewrite in another language for speed or deployment. \item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \endlist We introduce the Julia programming language and its design---a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience. }
 }
 
-@misc{srikrishnan_2022,
-	type = {preprint},
-	title = {Uncertainty analysis in multi-sector systems: {Considerations} for risk analysis, projection, and planning for complex systems},
-	copyright = {Creative Commons Attribution-NonCommercial 4.0 International (CC BY-NC 4.0)},
-	shorttitle = {Uncertainty analysis in multi-sector systems},
-	url = {http://www.essoar.org/doi/10.1002/essoar.10510113.2},
-	abstract = {Simulation models of multi-sector systems are increasingly used to understand societal resilience to climate and economic shocks and change. However, multi-sector systems are also subject to numerous},
-	language = {EN},
-	urldate = {2022-06-03},
-	journal = {Earth and Space Science Open Archive},
-	author = {Srikrishnan, Vivek and Lafferty, David C. and Wong, Tony E. and Lamontagne, Jonathan R. and Quinn, Julianne D. and Sharma, Sanjib and Molla, Nusrat J. and Herman, Jonathan D. and Sriver, Ryan L. and Morris, Jennifer and Lee, Ben Seiyon},
-	month = may,
-	year = {2022},
-	doi = {10.1002/essoar.10510113.2},
-	note = {Archive Location: world
-Publisher: Earth and Space Science Open Archive
-Section: Climatology (Global Change)},
-	file = {Snapshot:/Users/aewsma/Zotero/storage/QDNL89YJ/essoar.10510113.html:text/html},
+@article{srikrishnan_2022,
[…]
