### 2022-08-19 8b18a47
fixing a couple typos in paper, adding volume and year for Srikrishnan 2022 ref
