+++ joss_submission/paper.bib
@@ -22,15 +22,15 @@ eprint = {
 	abstract = {Abstract Simulation models of multi-sector systems are increasingly used to understand societal resilience to climate and economic shocks and change. However, multi-sector systems are also subject to numerous uncertainties that prevent the direct application of simulation models for prediction and planning, particularly when extrapolating past behavior to a nonstationary future. Recent studies have developed a combination of methods to characterize, attribute, and quantify these uncertainties for both single- and multi-sector systems. Here we review challenges and complications to the idealized goal of fully quantifying all uncertainties in a multi-sector model and their interactions with policy design as they emerge at different stages of analysis: (1) inference and model calibration; (2) projecting future outcomes; and (3) scenario discovery and identification of risk regimes. We also identify potential methods and research opportunities to help navigate the tradeoffs inherent in uncertainty analyses for complex systems. During this discussion, we provide a classification of uncertainty types and discuss model coupling frameworks to support interdisciplinary collaboration on multi-sector dynamics (MSD) research. Finally, we conclude with recommendations for best practices to ensure that MSD research can be properly contextualized with respect to the underlying uncertainties. This article is protected by copyright. All rights reserved.},
 	author = {Srikrishnan, Vivek and Lafferty, David C. and Wong, Tony E. and Lamontagne, Jonathan R. and Quinn, Julianne D. and Sharma, Sanjib and Molla, Nusrat J. and Herman, Jonathan D. and Sriver, Ryan L. and Morris, Jennifer F. and Lee, Ben Seiyon},
 	doi = {10.1029/2021EF002644},
-	eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2021EF002644},
+  year = {2022},
+  eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2021EF002644},
 	journal = {Earth's Future},
 	keywords = {multisector dynamics, uncertainty, calibration, projections, scenario discovery},
 	note = {e2021EF002644 2021EF002644},
-	number = {n/a},
 	pages = {e2021EF002644},
 	title = {Uncertainty analysis in multi-sector systems: Considerations for risk analysis, projection, and planning for complex systems},
 	url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2021EF002644},
-	volume = {n/a},
+	volume = {10},
 	bdsk-url-1 = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2021EF002644},
 	bdsk-url-2 = {https://doi.org/10.1029/2021EF002644}}
 }
