+++ joss_submission/paper.md
@@ -50,9 +50,9 @@ bibliography: paper.bib
 
 # Statement of need
 
-Assessing strategies to manage climate risks requires sound projections of future climate. For coastal risks, this includes projections of future sea levels. Major contributors to sea-level change include glaciers and ice caps, thermal expansion, land water storage, and the Greenland and Antarctic ice sheets. Characterizing coastal hazards and managing the associated risks requires resolving the tails of distributions. While many practitioners can obtain sea-level scenarios from syntheses reports, including the Intergovernmental Panel on Climate Change (IPCC, [@ipccar6ch9]), improving our understanding of the importance and impacts of the many inherent uncertainties often requires fast and efficient models to explore these uncertainties. Semi-empirical models for sea-level rise offer a computationally efficient method for studying uncertainties in future coastal hazards and fusing observational data with models [@Kopp2017; @Mengel2016; @Nauels2017; @Wong2017_brick]. These models can also be flexible and modular, which enables their use in integrated frameworks for assessing climate damages and examining the efficacy of climate risk management policies. Interfaces such as the Framework for Assessing Changes To Sea-level (`FACTS`, [@garner2022]) enable users to access these models and facilitate experimentation and intercomparisons. `FACTS`, coded primarily in Python, advances the older `LocalizeSL` MATLAB modules [@kopp_2014]. `FACTS` offers users a set of modules through which various forcing scenarios and modeling workflows can be combined to interrogate uncertainties in sea-level projections, particularly in structural uncertainties. `FACTS` is open source and freely available, and is a suitable source for local sea-level projections for many users.
+Assessing strategies to manage climate risks requires sound projections of future climate. For coastal risks, this includes projections of future sea levels. Major contributors to sea-level change include glaciers and ice caps, thermal expansion, land water storage, and the Greenland and Antarctic ice sheets. Characterizing coastal hazards and managing the associated risks requires resolving the tails of distributions. While many practitioners can obtain sea-level scenarios from synthesis reports, including the Intergovernmental Panel on Climate Change (IPCC, [@ipccar6ch9]), improving our understanding of the importance and impacts of the many inherent uncertainties often requires fast and efficient models to explore these uncertainties. Semi-empirical models for sea-level rise offer a computationally efficient method for studying uncertainties in future coastal hazards and fusing observational data with models [@Kopp2017; @Mengel2016; @Nauels2017; @Wong2017_brick]. These models can also be flexible and modular, which enables their use in integrated frameworks for assessing climate damages and examining the efficacy of climate risk management policies. Interfaces such as the Framework for Assessing Changes To Sea-level (`FACTS`, [@garner2022]) enable users to access these models and facilitate experimentation and intercomparisons. `FACTS`, coded primarily in Python, advances the older `LocalizeSL` MATLAB modules [@kopp_2014]. `FACTS` offers users a set of modules through which various forcing scenarios and modeling workflows can be combined to interrogate uncertainties in sea-level projections, particularly in structural uncertainties. `FACTS` is open source and freely available, and is a suitable source for local sea-level projections for many users.
 
[…]
